Extract RR Interval Data from a FIT File Object

Extracts RR interval data (HRV data) from a FIT file object. This function is designed for situations where you expect the FIT file to contain HRV data. If no HRV data is found, it issues a warning and returns an empty numeric vector.

Usage

extract_rr_data(fit_object)

Arguments

fit_object

A FIT file object, typically created by readFitFile. The object should be the result of reading a FIT file that is expected to contain HRV (RR interval) data.

Value

A data frame containing the RR data, obtained directly from hrv. If no HRV data is found, the function returns an empty numeric vector (numeric(0)). Unlike a previous version, this version does not perform any filtering on the returned HRV data.

Details

The function performs the following steps:

Input Validation: It validates the fit_object using a validate_fit_object function (assumed to be defined elsewhere in your package).

HRV Data Extraction: It attempts to extract HRV data using hrv.

HRV Data Check: It checks if the extracted HRV data is NULL or empty (has zero rows).

Return Value:

If HRV data is found and is not empty, the function returns the data frame returned by hrv.

If HRV data is NULL or empty, the function issues a warning message ("No HRV data found in FIT file") and returns an empty numeric vector (numeric(0)).

See also

Examples

if (FALSE) { # \dontrun{
# Assuming you have a FIT file named "activity.fit" that *should*
# contain HRV data.
fit_file <- FITfileR::readFitFile("activity.fit")
rr_data <- extract_rr_data(fit_file)

# If HRV data was present:
# rr_data will be a data.frame (the result of FITfileR::hrv()).

# If no HRV data was found:
# rr_data will be numeric(0).
} # }